requirements.txt
streamlit
opencv-python
Pillow
matplotlib
numpy
pandas
torch
torchvision
timm
segmentation-models-pytorch
imutils
scikit-image
albumentations
scikit-learn
tqdm
yolov5
paddleocr
ensemble-boxes
paddlepaddle
nltk
pytorch-lightning